requirements.txt
# Core dependencies
yfinance>=1.2.0
pandas>=2.0.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Monitoring
schedule>=1.2.0  # For scheduling runs

# Optional: For database storage
sqlite3  # Built-in with Python